refactor: enhance error handling and add auth module
