+++ src-tauri/src/auth/authentication.rs
